requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
websockets==12.0
python-dotenv==1.0.0
transformers==4.35.2
torch==2.1.0
accelerate==0.24.1
llama-cpp-python==0.2.23
huggingface_hub
numpy 